% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatuR.R
\name{AddNode}
\alias{AddNode}
\title{Add node to DB}
\usage{
AddNode(db, parent_node, name = "New_signature", reference = NA)
}
\arguments{
\item{db}{The database object to be updated}

\item{parent_node}{The parent node where the new node should be added}

\item{name}{Name for the new node}

\item{reference}{Optional text to describe the node}
}
\value{
An update database containing the new node
}
\description{
Add a new internal node to the DB. This generates a local, updated copy of the database.
}
\examples{
SignatuR <- AddNode(SignatuR, parent_node=SignatuR$Hs, name="New_category")
}
