% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatuR.R
\name{AddSignature}
\alias{AddSignature}
\title{Add signature to DB}
\usage{
AddSignature(db, node, name = "New_signature", signature = NA, reference = NA)
}
\arguments{
\item{db}{The database object to be updated}

\item{node}{A database node where the new signature should be added}

\item{name}{Signature name}

\item{signature}{Gene signature, as a vector of genes}

\item{reference}{A text describing source of the signature or other comments}
}
\value{
An update database containing the new signature
}
\description{
Add a new signature to the DB. This generates a local, updated copy of the database.
}
\examples{
SignatuR <- AddSignature(SignatuR, node=SignatuR$Mm$Cell_types,
    name="T_cell", reference="A simple T cell signature", signature=c("Cd2","Cd3d","Cd3e"))
}
