% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatuR.R
\name{csv2SignatuR}
\alias{csv2SignatuR}
\title{Convert from .csv to SignatuR format}
\usage{
csv2SignatuR(file = "mySignatuR.csv")
}
\arguments{
\item{file}{Source file (.csv)}
}
\value{
A SignatuR object
}
\description{
Read in a SignatuR DB in .csv format, and store in memory
}
\examples{
# Save DB
SignatuR2csv(SignatuR, file="mySignatuR.csv")
# Load it back
mySignatuR <- csv2SignatuR("mySignatuR.csv")
}
\seealso{
\link{SignatuR2csv}
}
