% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatuR.R
\name{GetSignature}
\alias{GetSignature}
\title{Get signatures from DB}
\usage{
GetSignature(node)
}
\arguments{
\item{node}{A database node, either internal or leaf}
}
\value{
A single signature or list of signatures
}
\description{
Return one or more signatures from a specified database node. If an
intermediate node is specified, all signatures below that node are returned
as a list of signatures.
}
\examples{
# Single signature
GetSignature(SignatuR$Mm$Programs$HeatShock)
# Multiple signature from intermediate node
s <- GetSignature(SignatuR$Mm$Programs)
lapply(s, head)
}
