% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatuR.R
\name{LoadSignatuR}
\alias{LoadSignatuR}
\title{Load SignatuR from local file}
\usage{
LoadSignatuR(file = "mySignatuR.csv")
}
\arguments{
\item{file}{Source file (.csv, .rds or .rda)}
}
\value{
A SignatuR object
}
\description{
Load a file storing a SignatuR object. See also the
function \link{SaveSignatuR} for saving existing DB.
}
\examples{
mySignatuR <- LoadSignatuR(file="mySignatuR.rds")
}
\seealso{
\link{SaveSignatuR}
}
