% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatuR.R
\name{RemoveSignature}
\alias{RemoveSignature}
\title{Remove signature from DB}
\usage{
RemoveSignature(node)
}
\arguments{
\item{node}{A database node to be removed}
}
\description{
Remove a given signature or node from the database
}
\examples{
SignatuR <- RemoveSignature(SignatuR, node=SignatuR$Hs$Compartments$TCR)
}
