% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatuR.R
\name{SaveSignatuR}
\alias{SaveSignatuR}
\title{Save a local copy of SignatuR}
\usage{
SaveSignatuR(db, file = "mySignatuR.csv")
}
\arguments{
\item{db}{The database object to be saved}

\item{file}{Destination file (.csv, .rds or .rda)}
}
\description{
Store a modified copy of your SignatuR DB, in different formats.
Saving to .csv allows editing of signatures in a text or
spreadsheet program
}
\examples{
# Save DB
SaveSignatuR(SignatuR, file="mySignatuR.csv")
# Load it back
mySignatuR <- LoadSignatuR("mySignatuR.csv")
}
\seealso{
\link{LoadSignatuR}
}
