% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatuR.R
\name{SignatuR2csv}
\alias{SignatuR2csv}
\title{Save a SignatuR to csv file}
\usage{
SignatuR2csv(db, file = "mySignatuR.csv")
}
\arguments{
\item{db}{The database object to be saved}

\item{file}{Destination file (.csv)}
}
\description{
Store a copy of your SignatuR DB to a readable .csv file
}
\examples{
# Save DB
SignatuR2csv(SignatuR, file="mySignatuR.csv")
# Load it back
mySignatuR <- csv2SignatuR("mySignatuR.csv")
}
\seealso{
\link{csv2SignatuR}
}
