% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatuR.R
\name{LoadSignatuR}
\alias{LoadSignatuR}
\title{Load SignatuR from local file}
\usage{
LoadSignatuR(file = "mySignatuR.rds")
}
\arguments{
\item{file}{Source file (.rds or .rda)}
}
\value{
A SignatuR object
}
\description{
Load a .rds or .rda file storing a SignatuR object
}
\examples{
mySignatuR <- LoadSignatuR(file="mySignatuR.rds")
}
\seealso{
\link{SaveSignatuR}
}
