% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatuR.R
\name{SaveSignatuR}
\alias{SaveSignatuR}
\title{Save a local copy of SignatuR}
\usage{
SaveSignatuR(db, file = "mySignatuR.rds")
}
\arguments{
\item{db}{The database object to be saved}

\item{file}{Destination file (.rds or .rda)}
}
\description{
Store a modified copy of your SignatuR DB, either to a .rds or .rda file
}
\examples{
# Save DB
SaveSignatuR(SignatuR, file="mySignatuR.rds")
# Load it back
mySignatuR <- LoadSignatuR("mySignatuR.rds")
}
\seealso{
\link{LoadSignatuR}
}
