% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatuR.R
\docType{data}
\name{SignatuR}
\alias{SignatuR}
\title{A database of gene signatures for single-cell data analysis}
\format{
A \code{data.tree} object with signatures organized by species and category. Gene lists are stored in the "Signature" attribute,
comments and references for the signature are stored in the "Reference" attribute.
\describe{
\item{See database structure}{\code{print(SignatuR)}}
\item{Plot database structure}{\code{library(DiagrammeR)} \cr
\code{plot(SignatuR)}}
\item{See database with annotations}{\code{print(SignatuR, "Reference","Signature")}}
\item{Extract a specific signature}{\code{SignatuR$Hs$Compartments$TCR$Signature}}
\item{Extract all signatures below a given node (Get accessor)}{\code{SignatuR$Hs$Compartments$Get("Signature", filterFun = isLeaf)}}
}
}
\usage{
SignatuR
}
\description{
The DB is structured in a tree format, using the \code{data.tree} data type. See the examples below to browse and access this data type.
print(SignatuR)
}
\keyword{datasets}
